+++ lib/features/astronomy_picture_of_the_day/data/datasources/astronomy_picture_remote_data_source.dart
@@ -3,11 +3,12 @@ import '../../domain/dtos/pagination.dart';
 
 /// Remote Data Source of Astronomy Pictures.
 abstract interface class AstronomyPictureRemoteDataSource {
-  /// Retrieves astronomy pictures according to the given date range, page and
-  /// number of pictures per page — [pagination].
+  /// Retrieves, in descending order of date, astronomy pictures according to
+  /// the given date date range, page and the number of pictures per page —
+  /// [pagination].
   ///
   /// Throws [DatasourceException] to indicate an operation error.
-  Future<AstronomyPicturesWithPagination> getAstronomyPictures(
+  Future<AstronomyPicturesWithPagination> getAstronomyPicturesDesc(
     Pagination pagination,
   );
 }